Import SQL changes

+++ Kahood2/src/main/java/Kahood2/Models/TeamStudent.java
@@ -28,7 +28,7 @@ public class TeamStudent {
         return teamId;
     }
 
-    public void setTeamId(int teamId) {
+    public void setTeamId(long teamId) {
         this.teamId = teamId;
     }
 
@@ -36,7 +36,7 @@ public class TeamStudent {
         return studentId;
     }
 
-    public void setStudentId(int studentId) {
+    public void setStudentId(long studentId) {
         this.studentId = studentId;
     }
 }
